For humble

+++ elevation_mapping_cupy/elevation_mapping_cupy/semantic_map.py
@@ -321,7 +321,7 @@ class SemanticMap:
         mer = mer.astype(cp.float32)
         mer = mer.view(dtype=cp.uint32)
         ma = cp.bitwise_and(mer, 0xFFFF, dtype=np.uint16)
-        ma = ma.view(np.float16)
+        ma = ma.view(np.float32)
         ma = ma.astype(np.float32)
         ind = cp.right_shift(mer, 16)
         return ma, ind
